backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
langchain==1.2.15
langchain-openai==1.1.12
llama-index-core==0.14.19
llama-index-embeddings-openai==0.5.2
llama-index-retrievers-bm25==0.7.1
openai==2.30.0
pypdf==6.10.2
python-dotenv==1.2.2
